New folder (2)/backend/requirements.txt.txt
Django==4.2
djangorestframework==3.14
psycopg2-binary==2.9